epdg: fix UE to UE traffic
UE to UE traffic when both UE are connected via the same ePDG
couldn't send traffic to each other because of xfrm policies.
The firewall on the ePDG was catching this traffic because it tried to
shortcut without passing the P-GW.

Use fwmark for both directions and mark them also strongswan which
will configure it to the xfrm policies.

Related: OS#6435